# specs/openapi.app.json
{
  "components": {
    "schemas": {
...
      "DemoInboundEmailRequest": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "numbers": {
            "items": {
              "type": "string"
            },
            "title": "Numbers",
            "type": "array"
          },
          "senderEmail": {
            "title": "Senderemail",
            "type": "string"
          }
        },
        "required": [
          "body",
          "numbers",
          "senderEmail"
        ],
        "title": "DemoInboundEmailRequest",
        "type": "object"
      },
...
      "EmailSender": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "senderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Senderid"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "email",
          "userId"
        ],
        "title": "EmailSender",
        "type": "object"
      },
      "EmailSenderRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "senderId": {
            "title": "Senderid",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "email",
          "senderId",
          "userId"
        ],
        "title": "EmailSenderRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "InboundRule": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/RuleAction"
          },
          "actionAddress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actionaddress"
          },
          "backupEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backupemail"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "keyword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword"
          },
          "matchKind": {
            "$ref": "#/components/schemas/MatchKind",
            "default": "ANY"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "ruleId": {
            "title": "Ruleid",
            "type": "string"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret"
          },
          "votes": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Votes",
            "type": "object"
          }
        },
        "required": [
          "action",
          "createdAt",
          "name",
          "ruleId"
        ],
        "title": "InboundRule",
        "type": "object"
      },
      "InboundRuleRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/RuleAction"
          },
          "actionAddress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actionaddress"
          },
          "backupEmail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backupemail"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "keyword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword"
          },
          "matchKind": {
            "$ref": "#/components/schemas/MatchKind",
            "default": "ANY"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          }
        },
        "required": [
          "action",
          "name"
        ],
        "title": "InboundRuleRequest",
        "type": "object"
      },
...
      "MatchKind": {
        "enum": [
          "ANY",
          "KEYWORD"
        ],
        "title": "MatchKind",
        "type": "string"
      },
...
      "RuleAction": {
        "enum": [
          "AUTO_REPLY",
          "EMAIL_FIXED",
          "EMAIL_USER",
          "GROUP_SMS",
          "MOVE_CONTACT",
          "POLL",
          "SEND_TO_MESSENGER",
          "SMS",
          "URL"
        ],
        "title": "RuleAction",
        "type": "string"
      },
...
      "Website": {
        "properties": {
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "registeredAt": {
            "title": "Registeredat",
            "type": "string"
          },
          "rejectedReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejectedreason"
          },
          "status": {
            "$ref": "#/components/schemas/WebsiteStatus"
          }
        },
        "required": [
          "domain",
          "registeredAt",
          "status"
        ],
        "title": "Website",
        "type": "object"
      },
      "WebsiteRejectionRequest": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "WebsiteRejectionRequest",
        "type": "object"
      },
      "WebsiteStatus": {
        "enum": [
          "APPROVED",
          "REJECTED",
          "UNDER_REVIEW"
        ],
        "title": "WebsiteStatus",
        "type": "string"
      },
      "WebsitesRequest": {
        "properties": {
          "domains": {
            "items": {
              "type": "string"
            },
            "title": "Domains",
            "type": "array"
          }
        },
        "required": [
          "domains"
        ],
        "title": "WebsitesRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "txtlocal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/app/demo/inbound-email": {
      "post": {
        "operationId": "demo_inbound_email_api_app_demo_inbound_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInboundEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Demo Inbound Email"
      }
    },
...
    "/api/app/email-senders": {
      "get": {
        "operationId": "list_email_senders_api_app_email_senders_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmailSender"
                  },
                  "title": "Response List Email Senders Api App Email Senders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "List Email Senders"
      },
      "post": {
        "operationId": "add_email_sender_api_app_email_senders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailSender"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Add Email Sender"
      }
    },
    "/api/app/email-senders/{id}": {
      "delete": {
        "operationId": "remove_email_sender_api_app_email_senders__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Remove Email Sender"
      }
    },
...
    "/api/app/rules/inbound": {
      "get": {
        "operationId": "list_inbound_rules_api_app_rules_inbound_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InboundRule"
                  },
                  "title": "Response List Inbound Rules Api App Rules Inbound Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "List Inbound Rules"
      },
      "post": {
        "operationId": "create_inbound_rule_api_app_rules_inbound_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundRule"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Create Inbound Rule"
      }
    },
    "/api/app/rules/inbound/{ruleId}": {
      "delete": {
        "operationId": "delete_inbound_rule_api_app_rules_inbound__ruleId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "ruleId",
            "required": true,
            "schema": {
              "title": "Ruleid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Delete Inbound Rule"
      },
      "put": {
        "operationId": "update_inbound_rule_api_app_rules_inbound__ruleId__put",
        "parameters": [
          {
            "in": "path",
            "name": "ruleId",
            "required": true,
            "schema": {
              "title": "Ruleid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundRule"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Update Inbound Rule"
      }
    },
...
    "/api/app/webhooks/{ruleId}/test": {
      "post": {
        "operationId": "test_webhook_api_app_webhooks__ruleId__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "ruleId",
            "required": true,
            "schema": {
              "title": "Ruleid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Test Webhook"
      }
    },
    "/api/app/websites": {
      "get": {
        "operationId": "list_websites_api_app_websites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Website"
                  },
                  "title": "Response List Websites Api App Websites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "List Websites"
      },
      "post": {
        "operationId": "register_websites_api_app_websites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsitesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Website"
                  },
                  "title": "Response Register Websites Api App Websites Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Register Websites"
      }
    },
...
    "/api/operator/accounts/{accountId}/websites/{domain}/approve": {
      "post": {
        "operationId": "approve_website_api_operator_accounts__accountId__websites__domain__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "title": "Accountid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Website"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Approve Website"
      }
    },
    "/api/operator/accounts/{accountId}/websites/{domain}/reject": {
      "post": {
        "operationId": "reject_website_api_operator_accounts__accountId__websites__domain__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "title": "Accountid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "title": "Domain",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteRejectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Website"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Reject Website"
      }
    },
...
  }
}